% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\docType{package}
\name{analytics_googleAuthR}
\alias{analytics_googleAuthR}
\alias{analytics_googleAuthR-package}
\title{Google Analytics API
Views and manages your Google Analytics data.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:24:42
filename: /Users/mark/dev/R/autoGoogleAPI/googleanalyticsv24.auto/R/analytics_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/analytics
\item https://www.googleapis.com/auth/analytics.readonly
}
}
