% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{data.get}
\alias{data.get}
\title{Returns Analytics report data for a view (profile).}
\usage{
data.get(ids, start.date, end.date, metrics, dimensions = NULL,
  filters = NULL, max.results = NULL, segment = NULL, sort = NULL,
  start.index = NULL)
}
\arguments{
\item{ids}{Unique table ID for retrieving report data}

\item{start.date}{Start date for fetching report data}

\item{end.date}{End date for fetching report data}

\item{metrics}{A comma-separated list of Analytics metrics}

\item{dimensions}{A comma-separated list of Analytics dimensions}

\item{filters}{A comma-separated list of dimension or metric filters to be applied to the report data}

\item{max.results}{The maximum number of entries to include in this feed}

\item{segment}{An Analytics advanced segment to be applied to the report data}

\item{sort}{A comma-separated list of dimensions or metrics that determine the sort order for the report data}

\item{start.index}{An index of the first entity to retrieve}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics
\item https://www.googleapis.com/auth/analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics, https://www.googleapis.com/auth/analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}
}
