% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_functions.R
\name{management.profiles.list}
\alias{management.profiles.list}
\title{Lists views (profiles) to which the user has access.}
\usage{
management.profiles.list(accountId, webPropertyId, max.results = NULL,
  start.index = NULL)
}
\arguments{
\item{accountId}{Account ID for the views (profiles) to retrieve}

\item{webPropertyId}{Web property ID for the views (profiles) to retrieve}

\item{max.results}{The maximum number of views (profiles) to include in this response}

\item{start.index}{An index of the first entity to retrieve}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics
\item https://www.googleapis.com/auth/analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics, https://www.googleapis.com/auth/analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/}{Google Documentation}
}

